ai-model-backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
python-dotenv
requests
scikit-learn